// frontend/src/api/model/githubCom4H1RZooraInternalDomainNotificationInboxItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */
import type { GithubCom4H1RZooraInternalDomainNotificationAudience } from './githubCom4H1RZooraInternalDomainNotificationAudience';
import type { GithubCom4H1RZooraInternalDomainNotificationCategory } from './githubCom4H1RZooraInternalDomainNotificationCategory';

export interface GithubCom4H1RZooraInternalDomainNotificationInboxItem {
  action_url?: string;
  audience?: GithubCom4H1RZooraInternalDomainNotificationAudience;
  body?: string;
  category?: GithubCom4H1RZooraInternalDomainNotificationCategory;
  created_at?: string;
  id?: string;
  organization_id?: string;
  read_at?: string;
  sender_id?: string;
  title?: string;
  updated_at?: string;
}