// frontend/src/api/model/githubCom4H1RZooraInternalDomainNotificationCategory.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */

export type GithubCom4H1RZooraInternalDomainNotificationCategory = typeof GithubCom4H1RZooraInternalDomainNotificationCategory[keyof typeof GithubCom4H1RZooraInternalDomainNotificationCategory];


export const GithubCom4H1RZooraInternalDomainNotificationCategory = {
  NotificationCategorySystem: 'system',
  NotificationCategoryOrg: 'org',
  NotificationCategoryClass: 'class',
  /** reserved for the future Bell feature */
  NotificationCategoryReminder: 'reminder',
} as const;